Move ConjureProductDependenciesExtension back to api package
